// src/app/types.d.ts
interface PowerInfo {
  hasGenerator: boolean;
  generators: Generator[];
  type: string;
  make: string;
  model: string;
  hours: number;
  powerPerEngine: number;
  fuel: string;
  count: number;
  range: number | null;
  cruiseSpeed: number | null;
  topSpeed: number;
}

interface Generator {
  generatorMake: string;
  generatorHours: number;
}

interface Photos {
  exterior: Media[];
  interior: Media[];
  deckLayout: Media[];
}

interface Accommodation {
  singleBeds: number;
  doubleBeds: number;
  cabins: number;
  showerWetHead: number;
  showerDryHead: number;
}

export interface TechnicalInfo {
  lengthOverall: number;
  lengthOfWaterline: number | null;
  beam: number;
  draft: number;
  dryWeight: number;
  displacementWeight: number | null;
  standardDraft: number | null;
  optionalDraft: number | null;
  keelWeight: number | null;
  mainSail: string | null;
  jib: string | null;
  spinnaker: string | null;
}